successfully uploads freestyle neo meter
