+++ app/reducers/devices.js
@@ -99,6 +99,13 @@ const devices = {
     source: {type: 'device', driverId: 'AbbottFreeStyleLibre'},
     enabled: {linux: true, mac: true, win: true}
   },
+  abbottfreestyleneo: {
+    instructions: 'Plug in meter with micro-USB cable',
+    key: 'abbottfreestyleneo',
+    name: 'Abbott FreeStyle Precision/Optium Neo',
+    source: {type: 'device', driverId: 'AbbottFreeStyleNeo'},
+    enabled: {linux: true, mac: true, win: true}
+  },
   bayercontournext: {
     instructions: 'Plug meter into USB port',
     key: 'bayercontournext',
@@ -147,13 +154,6 @@ const devices = {
     key: 'truemetrix',
     source: {type: 'device', driverId: 'TrueMetrix'},
     enabled: {mac: true, win: true, linux: true}
-  },
-  abbottprecisionneo: {
-    instructions: 'Plug in meter with micro-USB cable',
-    key: 'abbottprecisionneo',
-    name: 'Abbott FreeStyle Precision/Optium Neo',
-    source: {type: 'device', driverId: 'AbbottPrecisionNeo'},
-    enabled: {linux: true, mac: true, win: true}
   },
 };
