+++ lib/core/device.js
@@ -43,6 +43,7 @@ var oneTouchVerio = require('../drivers/onetouch/oneTouchVerio');
 var oneTouchVerioIQ = require('../drivers/onetouch/oneTouchVerioIQ');
 var abbottFreeStyleLite = require('../drivers/abbott/abbottFreeStyleLite');
 var abbottFreeStyleLibre = require('../drivers/abbott/abbottFreeStyleLibre');
+var abbottFreeStyleNeo = require('../drivers/abbott/abbottFreeStyleNeo');
 var bayerContourNext = require('../drivers/bayer/bayerContourNext');
 var animasDriver = require('../drivers/animas/animasDriver');
 var medtronicDriver = require('../drivers/medtronic/medtronicDriver');
@@ -50,7 +51,6 @@ var medtronic600Driver = require('../drivers/medtronic600/medtronic600Driver');
 var TrueMetrixDriver = require('../drivers/trividia/trueMetrix');
 var accuChekUSBDriver = require('../drivers/roche/accuChekUSB');
 var bluetoothLEDriver = require('../drivers/bluetoothLE/bluetoothLEDriver');
-var PrecisionNeoDriver = require('../drivers/abbott/abbottPrecisionNeo');
 
 var device = {
   log: require('bows')('Device')
@@ -73,6 +73,7 @@ device._deviceDrivers = {
   'OneTouchVerioIQ': oneTouchVerioIQ,
   'AbbottFreeStyleLite': abbottFreeStyleLite,
   'AbbottFreeStyleLibre': abbottFreeStyleLibre,
+  'AbbottFreeStyleNeo': abbottFreeStyleNeo,
   'BayerContourNext': bayerContourNext,
   'Animas': animasDriver,
   'Medtronic': medtronicDriver,
@@ -80,7 +81,6 @@ device._deviceDrivers = {
   'TrueMetrix': TrueMetrixDriver,
   'AccuChekUSB': accuChekUSBDriver,
   'BluetoothLE': bluetoothLEDriver,
-  'AbbottPrecisionNeo': PrecisionNeoDriver,
 };
 
 device._deviceComms = {
@@ -92,6 +92,7 @@ device._deviceComms = {
   'OneTouchVerioIQ': serialDevice,
   'AbbottFreeStyleLite': serialDevice,
   'AbbottFreeStyleLibre': hidDevice,
+  'AbbottFreeStyleNeo': hidDevice,
   'Tandem': serialDevice,
   'BayerContourNext': hidDevice,
   'Animas': serialDevice,
@@ -100,7 +101,6 @@ device._deviceComms = {
   'TrueMetrix': hidDevice,
   'AccuChekUSB': usbDevice,
   'BluetoothLE': bleDevice,
-  'AbbottPrecisionNeo': hidDevice,
 };
 
 device._driverManifests = {
@@ -159,6 +159,12 @@ device._driverManifests = {
     usb: [
       {vendorId: 6753, productId: 13904}, // FreeStyle Libre
       {vendorId: 6753, productId: 13936}  // FreeStyle Libre Pro
+    ]
+  },
+  'AbbottFreeStyleNeo': {
+    mode: 'HID',
+    usb: [
+      {vendorId: 6753, productId: 14416}, // FreeStyle Optium Neo
     ]
   },
   'BayerContourNext': {
@@ -231,12 +237,6 @@ device._driverManifests = {
   },
   'BluetoothLE': {
     mode: 'bluetooth',
-  },
-  'AbbottPrecisionNeo': {
-    mode: 'HID',
-    usb: [
-      {vendorId: 6753, productId: 14416}, // FreeStyle Optium Neo
-    ]
   },
 };
