+++ lib/drivers/abbott/freeStyleLibreProtocol.js
@@ -610,13 +610,13 @@ export default class FreeStyleLibreProtocol {
         return errorCallback(new Error('Invalid database record checksum.'), data);
       }
 
-      dbRecords = dbRecords.split('\r\n');
+      dbRecords = dbRecords.split('\r\n').filter(Boolean);
 
       if (dbRecordCount !== dbRecords.length) {
         return errorCallback(new Error(`Invalid database record count: ${dbRecordCount} != ${dbRecords.length}`), data);
       }
 
-      return successCallback(null, dbRecords);
+      return successCallback(dbRecords);
     }, errorCallback);
   }
